LayoutTests/fast/mediastream/getDisplayMedia-max-constraints.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <title>getDisplayMedia track support of max constraints</title>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>

        <script>
            if (window.internals)
                window.internals.settings.setScreenCaptureEnabled(true);

            async function callGetDisplayMedia(options)
            {
                let promise;
                window.internals.withUserGesture(() => {
                    promise = navigator.mediaDevices.getDisplayMedia(options);
                });
                const stream = await promise; 
                const track = stream.getVideoTracks()[0];

                // getSettings is not computing the correct parameters right away. We should probably fix this.
                while (true) {
                    const settings = track.getSettings();
                    if (settings.width && settings.height)
                        break;
                    await new Promise(resolve => setTimeout(resolve, 50));
                }
                return stream;
            }

            // getSettings is not computing the correct parameters right away. We should probably fix this.
            async function waitForHeight(track, value) {
                while (true) {
                    const settings = track.getSettings();
                    if (settings.height === value)
                        break;
                    await new Promise(resolve => setTimeout(resolve, 50));
                }
            }

            // getSettings is not computing the correct parameters right away. We should probably fix this.
            async function waitForWidth(track, value) {
                while (true) {
                    const settings = track.getSettings();
                    if (settings.width === value)
                        break;
                    await new Promise(resolve => setTimeout(resolve, 50));
                }
            }

            let defaultWidth;
            let defaultHeight;
            promise_test(async () => {
                stream = await callGetDisplayMedia({ video: true });
                let settings = stream.getVideoTracks()[0].getSettings();
                defaultWidth = settings.width;
                defaultHeight = settings.height;
            }, "setup");
            
            promise_test(async (test) => {
                const stream = await callGetDisplayMedia({ video: { height: { max: 500 }, width : { max : 500 } } });

                const expectedHeight =Math.floor(500 * (defaultHeight / defaultWidth)) 
                await waitForHeight(stream.getVideoTracks()[0], expectedHeight);

                const settings = stream.getVideoTracks()[0].getSettings()
                assert_equals(settings.width, 500);
            }, "Maximize the width if max height is too big");

            promise_test(async (test) => {
                const stream = await callGetDisplayMedia({ video: { height: { max: 100 }, width : { max : 500 } } });

                const expectedWidth = Math.floor(100 * (defaultWidth / defaultHeight)) 
                await waitForWidth(stream.getVideoTracks()[0], expectedWidth);

                const settings = stream.getVideoTracks()[0].getSettings()
                assert_equals(settings.height, 100);
            }, "Maximize the height if max width is too big");

            promise_test(async (test) => {
                const stream = await callGetDisplayMedia({ video: { height: { max: 500000 }, width : { max : 500000 } } });

                await waitForHeight(stream.getVideoTracks()[0], defaultHeight);

                const settings = stream.getVideoTracks()[0].getSettings()
                assert_equals(settings.width, defaultWidth);
            }, "No effect of the max values if they are too big");
        </script>
    </body>
</html>
